add comment for api bug, reorg ega services
